Add student removal functionality
Added a new method in the School class that allows removal of students by their information. Also, an error message will be thrown if the student was not found in the school list. The Main class now includes calling this new method to remove students and then printing the student list. Adjustments to the printStudents method have also been made to account for cases when there are no students.